+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # testclusters-go Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -8,11 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
-- add `cluster.*K3dCluster.WriteKubeConfig()` to produce a KUBECONFIG for cluster debugging
+- add `cluster.*K3dCluster.WriteKubeConfig()` to produce a KUBECONFIG for manual cluster debugging
 - [#16] provide feedback on node health for more cluster transparency
 - add convenience constructor `cluster.NewK3dClusterWithOpts()` for further cluster customizing
-  - f. i. log level or custom cluster prefix
+   - f. i. log level or custom cluster prefix
+- [#16] make disk pressure condition configurable
+   - see also the [feature docs](docs/features.md#configure-hard-eviction-string-for-kubelet)
 
 ## Changed
+
 - [#16] clean up cluster containers more robustly
-  - containers may still prevail cleaning if the cluster test will be hard-terminated (f. i. pressing the Debug-Kill 💀 button in IntelliJ IDEA)
+   - containers may still prevail cleaning if the cluster test will be hard-terminated (f. i. pressing the Debug-Kill 💀
+     button in IntelliJ IDEA)